+++ bugsnag-android-core/src/test/java/com/bugsnag/android/BreadcrumbFacadeTest.java
@@ -75,4 +75,9 @@ public class BreadcrumbFacadeTest {
     public void dateValid() {
         assertEquals(new Date(0).getTime(), crumb.getTimestamp().getTime());
     }
+
+    @Test
+    public void stringDateValid() {
+        assertEquals("1970-01-01T00:00:00.000Z", crumb.getStringTimestamp());
+    }
 }