+++ bugsnag-android-core/src/test/java/com/bugsnag/android/PluginClientTest.kt
@@ -1,5 +1,7 @@
 package com.bugsnag.android
 
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNull
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mock
@@ -23,5 +25,12 @@ class PluginClientTest {
         val pluginClient = PluginClient(setOf(plugin), config, NoopLogger)
         pluginClient.loadPlugins(client)
         Mockito.verify(plugin, times(1)).load(client)
+    }
+
+    @Test
+    fun findPlugin() {
+        val pluginClient = PluginClient(setOf(plugin), config, NoopLogger)
+        assertNull(pluginClient.findPlugin(String::class.java))
+        assertEquals(plugin, pluginClient.findPlugin(plugin::class.java))
     }
 }